Move CI build to Windows, disable NetFx tests on Mono

+++ tools/builder/targets/TestFx.cs
@@ -1,3 +1,4 @@
+using System;
 using System.IO;
 using System.Threading.Tasks;
 using Xunit.BuildTools.Models;
@@ -13,6 +14,14 @@ public static class TestFx
 	public static Task OnExecute(BuildContext context)
 	{
 		context.BuildStep("Running .NET Framework tests");
+
+		if (context.NeedMono)
+		{
+			context.WriteLineColor(ConsoleColor.Yellow, "Skipping .NET Framework tests on non-Windows OSes.");
+			Console.WriteLine();
+
+			return Task.CompletedTask;
+		}
 
 		var resultPath = Path.Combine(context.BaseFolder, "artifacts", "test");
 		File.Delete(Path.Combine(resultPath, "netfx.trx"));
